Refactored population in area analysis
