+++ lib/node/nodes/point-in-polygon/point-in-polygon-query-builder.js
@@ -2,11 +2,11 @@
 
 var fs = require('fs');
 var dot = require('dot');
-var pointInPolygon = fs.readFileSync(__dirname + '/point-in-polygon.sql');
+var pointInPolygonQuery = fs.readFileSync(__dirname + '/point-in-polygon.sql');
 
 dot.templateSettings.strip = false;
 
-var pointInPolygonTemplate = dot.template(pointInPolygon);
+var pointInPolygonTemplate = dot.template(pointInPolygonQuery);
 
 module.exports = function buildQuery(it) {
     it.pointsColumns = it.columnNames.map(function(name) {
